+++ substratevm/src/com.oracle.svm.agent/src/com/oracle/svm/agent/NativeImageAgent.java
@@ -38,7 +38,6 @@ import java.nio.file.attribute.FileTime;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.ConcurrentModificationException;
 import java.util.Date;
 import java.util.List;
@@ -48,7 +47,6 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
-import java.util.regex.Pattern;
 
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.ProcessProperties;
@@ -76,12 +74,10 @@ import com.oracle.svm.configure.filters.FilterConfigurationParser;
 import com.oracle.svm.configure.filters.HierarchyFilterNode;
 import com.oracle.svm.configure.trace.AccessAdvisor;
 import com.oracle.svm.configure.trace.TraceProcessor;
-import com.oracle.svm.core.SubstrateUtil;
 import com.oracle.svm.core.configure.ConfigurationFile;
 import com.oracle.svm.core.jni.headers.JNIEnvironment;
 import com.oracle.svm.core.jni.headers.JNIJavaVM;
 import com.oracle.svm.core.jni.headers.JNIObjectHandle;
-import com.oracle.svm.driver.NativeImage;
 import com.oracle.svm.driver.metainf.NativeImageMetaInfWalker;
 import com.oracle.svm.jvmtiagentbase.JNIHandleSet;
 import com.oracle.svm.jvmtiagentbase.JvmtiAgentBase;
@@ -143,7 +139,6 @@ public final class NativeImageAgent extends JvmtiAgentBase<NativeImageAgentJNIHa
         boolean experimentalClassDefineSupport = false;
         boolean experimentalUnsafeAllocationSupport = false;
         boolean experimentalOmitClasspathConfig = false;
-        boolean build = false;
         boolean configurationWithOrigins = false;
         List<String> conditionalConfigUserPackageFilterFiles = new ArrayList<>();
         List<String> conditionalConfigClassNameFilterFiles = new ArrayList<>();
@@ -206,8 +201,6 @@ public final class NativeImageAgent extends JvmtiAgentBase<NativeImageAgentJNIHa
                 if (configWritePeriodInitialDelay < 0) {
                     return usage(1, "config-write-initial-delay-secs must be an integer greater or equal to 0");
                 }
-            } else if (isBooleanOption(token, "build")) {
-                build = getBooleanTokenValue(token);
             } else if (isBooleanOption(token, "experimental-configuration-with-origins")) {
                 configurationWithOrigins = getBooleanTokenValue(token);
             } else if (token.startsWith("experimental-conditional-config-filter-file=")) {
@@ -223,9 +216,9 @@ public final class NativeImageAgent extends JvmtiAgentBase<NativeImageAgentJNIHa
             }
         }
 
-        if (traceOutputFile == null && configOutputDir == null && !build) {
+        if (traceOutputFile == null && configOutputDir == null) {
             configOutputDir = transformPath(AGENT_NAME + "_config-pid{pid}-{datetime}/");
-            inform("no output/build options provided, tracking dynamic accesses and writing configuration to directory: " + configOutputDir);
+            inform("no output options provided, tracking dynamic accesses and writing configuration to directory: " + configOutputDir);
         }
 
         if (configurationWithOrigins && !conditionalConfigUserPackageFilterFiles.isEmpty()) {
@@ -377,14 +370,6 @@ public final class NativeImageAgent extends JvmtiAgentBase<NativeImageAgentJNIHa
             }
         }
 
-        if (build) {
-            int status = buildImage(jvmti);
-            if (status == 0) {
-                System.exit(status);
-            }
-            return status;
-        }
-
         try {
             BreakpointInterceptor.onLoad(jvmti, callbacks, tracer, this, interceptedStateSupplier,
                             experimentalClassLoaderSupport, experimentalClassDefineSupport, experimentalUnsafeAllocationSupport, trackReflectionMetadata);
@@ -499,49 +484,6 @@ public final class NativeImageAgent extends JvmtiAgentBase<NativeImageAgentJNIHa
                 warn("Failed to walk the classpath entry: " + cpEntry + " Reason: " + e);
             }
         }
-    }
-
-    private static final Pattern propertyBlacklist = Pattern.compile("(java\\..*)|(sun\\..*)|(jvmci\\..*)");
-    private static final Pattern propertyWhitelist = Pattern.compile("(java\\.library\\.path)|(java\\.io\\.tmpdir)");
-
-    private static int buildImage(JvmtiEnv jvmti) {
-        System.out.println("Building native image ...");
-        String classpath = Support.getSystemProperty(jvmti, "java.class.path");
-        if (classpath == null) {
-            return usage(1, "Build mode could not determine classpath.");
-        }
-        String javaCommand = Support.getSystemProperty(jvmti, "sun.java.command");
-        String mainClassMissing = "Build mode could not determine main class.";
-        if (javaCommand == null) {
-            return usage(1, mainClassMissing);
-        }
-        String mainClass = SubstrateUtil.split(javaCommand, " ")[0];
-        if (mainClass.isEmpty()) {
-            return usage(1, mainClassMissing);
-        }
-        List<String> buildArgs = new ArrayList<>();
-        // buildArgs.add("--verbose");
-        String[] keys = Support.getSystemProperties(jvmti);
-        for (String key : keys) {
-            boolean whitelisted = propertyWhitelist.matcher(key).matches();
-            boolean blacklisted = !whitelisted && propertyBlacklist.matcher(key).matches();
-            if (blacklisted) {
-                continue;
-            }
-            buildArgs.add("-D" + key + "=" + Support.getSystemProperty(jvmti, key));
-        }
-        if (mainClass.toLowerCase().endsWith(".jar")) {
-            buildArgs.add("-jar");
-        } else {
-            buildArgs.addAll(Arrays.asList("-cp", classpath));
-        }
-        buildArgs.add(mainClass);
-        buildArgs.add(AGENT_NAME + ".build");
-        // System.out.println(String.join("\n", buildArgs));
-        Path javaHome = Paths.get(Support.getSystemProperty(jvmti, "java.home"));
-        String userDirStr = Support.getSystemProperty(jvmti, "user.dir");
-        NativeImage.agentBuild(javaHome, userDirStr == null ? null : Paths.get(userDirStr), buildArgs);
-        return 0;
     }
 
     private static String transformPath(String path) {